web_scraping/targeted_language_scraper/requirements.txt
# Core Framework
scrapy>=2.13.0,<2.14.0

# Web Scraping & JS Rendering
scrapy-playwright>=0.0.44,<0.1.0
playwright>=1.56.0,<2.0.0

# Config Generator
requests>=2.32.3,<2.33.0
beautifulsoup4>=4.14.0,<4.15.0
lxml

# PDF & Language Detection
PyPDF2>=3.0.0,<4.0.0
fasttext-wheel>=0.9.2,<1.0.0
huggingface-hub>=1.0.0,<2.0.0